rtables.officer: Exporting Tools for 'rtables'

Designed to create and display complex tables with R, the 'rtables' R package allows cells in an 'rtables' object to contain any high-dimensional data structure, which can then be displayed with cell-specific formatting instructions. Additionally, the 'rtables.officer' package supports export formats related to the Microsoft Office software suite, including Microsoft Word ('docx') and Microsoft PowerPoint ('pptx').

See also

Author

Maintainer: Joe Zhu joe.zhu@roche.com (ORCID)

Authors:

Davide Garolini davide.garolini@roche.com (ORCID)

Emily de la Rua emily.de_la_rua@contractors.roche.com (ORCID)

Abinaya Yogasekaram abinaya.yogasekaram@contractors.roche.com (ORCID)

Other contributors:

Gabriel Becker gabembecker@gmail.com [contributor]

F. Hoffmann-La Roche AG [copyright holder, funder]